tasklist finish

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,17 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
+        //$tasks = Task::paginate(10);
+    if (\Auth::check()){  
+        $user = \Auth::user();
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
         return view('tasks.index',[
             'tasks' => $tasks,
             ]);
+    }else{
+        return view('welcome');
+    }
     }
 
     /**
@@ -48,12 +54,12 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
+     
         return redirect('/');
     }
 
@@ -66,10 +72,12 @@ class TasksController extends Controller
     public function show($id)
     {
          $task = Task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -81,10 +89,12 @@ class TasksController extends Controller
     public function edit($id)
     {
          $task = Task::find($id);
-
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -101,10 +111,11 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         $task = Task::find($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
+        if (\Auth::id() === $task->user_id) {
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+        }
         return redirect('/');
     }
 
@@ -116,9 +127,10 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-         $task = Task::find($id);
-        $task->delete();
-
+        $task = Task::find($id);
+        if (\Auth::id() === $task->user_id) { 
+            $task->delete();
+        }
         return redirect('/');
     }
 }
